流星アニメーションを Web Animations API で実装する
